MNISTService/requirements.txt
fastapi~=0.72.0
numpy~=1.19.5
Pillow~=8.0.1
tensorflow~=2.7.0
uvicorn~=0.17.0
pydantic~=1.9.0
python-multipart~=0.0.5